requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
jinja2
python-multipart
pydantic
pydantic-settings
httpx
openai
pgvector
beautifulsoup4
lxml
python-slugify
pypdf
pdfplumber
itsdangerous
passlib[bcrypt]
alembic
gunicorn
pymupdf
pillow
pytesseract